ffz() takes ints, so change the file_map to being int based

+++ filesetup.c
@@ -601,7 +601,7 @@ int init_random_map(struct thread_data *td)
 				(unsigned long long) td->o.rw_min_bs;
 		num_maps = (blocks + BLOCKS_PER_MAP - 1) /
 				(unsigned long long) BLOCKS_PER_MAP;
-		f->file_map = smalloc(num_maps * sizeof(long));
+		f->file_map = smalloc(num_maps * sizeof(int));
 		if (f->file_map) {
 			f->num_maps = num_maps;
 			continue;
